[#1742] Move ROS header struct into common crate
