+++ .just/publish.just
@@ -117,6 +117,7 @@ INTEGRATIONS_ZENOH_CRATES_TO_IGNORE := `echo \
     iceoryx2-gateway-conformance-tests`
 
 INTEGRATIONS_ROS2_CRATES_TO_PUBLISH := `echo \
+    iceoryx2-integrations-ros2-interop         \
     iceoryx2-integrations-ros2-gateway-backend`
 
 INTEGRATIONS_ROS2_CRATES_TO_IGNORE := `echo \
