+++ integrations/ros2/gateway-backend/Cargo.toml
@@ -15,6 +15,7 @@ iceoryx2 = { workspace = true }
 iceoryx2-bb-concurrency = { workspace = true }
 iceoryx2-log = { workspace = true }
 iceoryx2-gateway-backend = { workspace = true }
+iceoryx2-integrations-ros2-interop = { workspace = true }
 
 libloading = { workspace = true }
 r2r_rcl = { workspace = true }
