+++ integrations/ros2/interop/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "iceoryx2-integrations-ros2-interop"
+description = "iceoryx2-integrations: interfaces for interoperation with ros 2"
+categories = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+keywords = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+version = { workspace = true }
+
+[dependencies]
+iceoryx2 = { workspace = true }
