+++ integrations/ros2/interop/README.md
@@ -1,0 +1,10 @@
+# iceoryx2-integrations-ros2-interop
+
+> [!IMPORTANT]
+> The ROS 2 integrations are currently prototypes and require validation in
+> real ROS 2 workflows. Only recommended for experiementation in development
+> deployments.
+>
+> If encountering issues, create an issue to help us converge to stability.
+
+Interfaces for interoperation with ROS 2.
