+++ integrations/ros2/workspace/src/demo_nodes/src/bin/prefix_mapping_plain_struct_translator_subscriber.rs
@@ -26,8 +26,9 @@
 
 use core::time::Duration;
 
-use demo_nodes_iceoryx2::{RosHeader, Twist};
+use demo_nodes_iceoryx2::Twist;
 use iceoryx2::prelude::*;
+use iceoryx2_integrations_ros2_interop::RosHeader;
 
 /// The iceoryx2 service mapped by the name prefix to the ROS 2 topic
 /// `/cmd_vel`.
