+++ integrations/ros2/workspace/src/demo_nodes/src/bin/static_mapping_plain_struct_translator_subscriber.rs
@@ -27,7 +27,8 @@
 use core::time::Duration;
 
 use iceoryx2::prelude::*;
-use demo_nodes_iceoryx2::{RosHeader, Twist};
+use demo_nodes_iceoryx2::Twist;
+use iceoryx2_integrations_ros2_interop::RosHeader;
 
 /// The iceoryx2 service paired with the ROS 2 topic `/cmd_vel` in
 /// `static_mapping_cmdvel.toml`.
